Update blinkt-overlay.dts
Provide a more descriptive node name for the SPI bus so udev rules can easily match on it.
Specify that the device number assigned to the spidev device file in /dev may not always be three. It depends on the order
the SPI buses are discovered, among other magical things the kernel does that us mortals do not fully understand.
Tidy up the empty cells - specify an empty group of cells instead of a single cell evaluating to zero.